[BTAT-2118 ] removed old IntegrationTestConstants and refactored IncomeSourceIntegrationTestConstants
